+++ client-elastic4s/src/main/scala/com/klibisz/elastiknn/client/ElastiknnRequests.scala
@@ -1,5 +1,6 @@
 package com.klibisz.elastiknn.client
 
+import com.klibisz.elastiknn.ELASTIKNN_NAME
 import com.klibisz.elastiknn.api.{ElasticsearchCodec, Mapping, NearestNeighborsQuery, Vec}
 import com.sksamuel.elastic4s.{ElasticDsl, Indexes, XContentBuilder, XContentFactory}
 import com.sksamuel.elastic4s.requests.indexes.IndexRequest
@@ -7,36 +8,76 @@ import com.sksamuel.elastic4s.requests.mappings.PutMappingRequest
 import com.sksamuel.elastic4s.requests.searches.SearchRequest
 import com.sksamuel.elastic4s.requests.searches.queries.CustomQuery
 
+/**
+  * Methods for creating Elastic4s requests for common elastiknn tasks.
+  * Methods are optimized for documents containing only a vector and running searches as quickly as possible.
+  * For example, we store the ID as a doc value instead of using the default stored ID which is slower to access.
+  * I am open to creating less-optimized, more-convenient methods in the future.
+  */
 trait ElastiknnRequests {
 
-  def indexVec(indexName: String, fieldName: String, vec: Vec, id: Option[String] = None): IndexRequest = {
-    val xcb = XContentFactory.jsonBuilder.rawField(fieldName, ElasticsearchCodec.nospaces(vec))
-    IndexRequest(indexName, source = Some(xcb.string()), id = id)
+  /**
+    * Create a request for indexing a vector.
+    *
+    * @param index Name of the index.
+    * @param vecField Field where vector is stored.
+    * @param vec Vector to index.
+    * @param storedIdField Field where document ID is stored.
+    * @param id Document ID. Stored as the ID known by Elasticsearch, and in the document for faster retrieval.
+    * @return Instance of a [[com.sksamuel.elastic4s.requests.indexes.IndexRequest]].
+    */
+  def index(index: String, vecField: String, vec: Vec, storedIdField: String, id: String): IndexRequest = {
+    val xcb = XContentFactory.jsonBuilder.rawField(vecField, ElasticsearchCodec.nospaces(vec)).field(storedIdField, id)
+    IndexRequest(index, source = Some(xcb.string()), id = Some(id))
   }
 
-  def nearestNeighborsQuery(index: String,
-                            query: NearestNeighborsQuery,
-                            k: Int,
-                            fetchSource: Boolean = false,
-                            preference: Option[String] = None): SearchRequest = {
+  /**
+    * Create a request for running a nearest neighbors query.
+    * Optimized for high performance, so it returns the document ID in the body.
+    * Sets the preference parameter (see: https://www.elastic.co/guide/en/elasticsearch/reference/master/consistent-scoring.html)
+    *  as the hash of the query for more deterministic results.
+    *
+    * @param index Index being searched against.
+    * @param query Constructed query, containing the vector, field, etc.
+    * @param k Number of results to return.
+    * @param storedIdField Field containing the document ID. See [[ElastiknnRequests.index()]] method.
+    * @return Instance of [[com.sksamuel.elastic4s.requests.searches.SearchRequest]].
+    */
+  def nearestNeighbors(index: String, query: NearestNeighborsQuery, k: Int, storedIdField: String): SearchRequest = {
     val json = ElasticsearchCodec.nospaces(query)
     val customQuery = new CustomQuery {
-      override def buildQueryBody(): XContentBuilder = XContentFactory.jsonBuilder.rawField("elastiknn_nearest_neighbors", json)
+      override def buildQueryBody(): XContentBuilder =
+        XContentFactory.jsonBuilder.rawField(s"${ELASTIKNN_NAME}_nearest_neighbors", json)
     }
-    val request = ElasticDsl.search(index).query(customQuery).fetchSource(fetchSource).size(k)
-    // https://www.elastic.co/guide/en/elasticsearch/reference/master/consistent-scoring.html
-    preference match {
-      case Some(pref) => request.preference(pref)
-      case None       => request
-    }
+    ElasticDsl
+      .search(index)
+      .query(customQuery)
+      .fetchSource(false)
+      .storedFields("_none_")
+      .docValues(Seq(storedIdField))
+      .preference(query.hashCode.toString)
+      .size(k)
   }
 
-  def putMapping(index: String, field: String, mapping: Mapping): PutMappingRequest = {
+  /**
+    * Create a mapping containing a vector field and a stored ID field.
+    *
+    * @param index Index to which this mapping is applied.
+    * @param vecField Field where vector is stored.
+    * @param storedIdField Field where ID is stored.
+    * @param vecMapping Mapping for the stored vector.
+    * @return Instance of [[com.sksamuel.elastic4s.requests.mappings.PutMappingRequest]].
+    */
+  def putMapping(index: String, vecField: String, storedIdField: String, vecMapping: Mapping): PutMappingRequest = {
     val mappingJsonString =
       s"""
          |{
          |  "properties": {
-         |    "$field": ${ElasticsearchCodec.encode(mapping).spaces2}
+         |    "$vecField": ${ElasticsearchCodec.encode(vecMapping).spaces2},
+         |    "$storedIdField": {
+         |      "type": "keyword",
+         |      "store": true
+         |    }
          |  }
          |}
          |""".stripMargin