+++ plugin/src/main/scala/com/klibisz/elastiknn/query/LshQuery.scala
@@ -23,13 +23,10 @@ object LshQuery {
     val terms = lshFunction(query).map(h => new BytesRef(UnsafeSerialization.writeInt(h)))
 
     val scoreFunction = (lrc: LeafReaderContext) => {
-      val binaryDocValues = lrc.reader.getBinaryDocValues(ExactQuery.vectorDocValuesField(field))
-      (docId: Int, _: Int) =>
-        if (binaryDocValues.advanceExact(docId)) {
-          val bref = binaryDocValues.binaryValue()
-          val storedVec = codec.decode(bref.bytes, bref.offset, bref.length)
-          lshFunction.exact(query, storedVec)
-        } else throw new RuntimeException(s"Couldn't advance to doc with id [$docId]")
+      val cachedReader = new ExactQuery.StoredVecReader[S](lrc, field)
+      (docId: Int, approximateScore: Int) =>
+        val storedVec = cachedReader(docId)
+        lshFunction.exact(query, storedVec)
     }
 
     new MatchTermsAndScoreQuery(